+++ packages/fe-container/src/components/side-menu/SideMenu.test.tsx
@@ -5,10 +5,10 @@ import userEvent from '@testing-library/user-event'
 
 import {SideMenu} from './SideMenu'
 import RenderWithReactIntl from '../../__tests__/utils'
-import {history, isPluginLoaded, registerPlugin} from '@plugins/PluginsLoaderFacade'
+import {history, isPluginLoaded, registerPlugin} from '@plugins-utils/PluginsLoaderFacade'
 
-jest.mock('../../plugins/PluginsLoaderFacade', () => ({
-  ...jest.requireActual('../../plugins/PluginsLoaderFacade'),
+jest.mock('@plugins-utils/PluginsLoaderFacade', () => ({
+  ...jest.requireActual('@plugins-utils/PluginsLoaderFacade'),
   isPluginLoaded: jest.fn(),
   history: {
     listen: jest.fn(() => {
