+++ packages/fe-container/src/plugins-utils/PluginsLoaderFacade.ts
